requirements.txt
fastapi
uvicorn
streamlit
requests
httpx
python-slugify
python-multipart
openai
langchain
pyttsx3
soundfile
numpy==1.24.3
pandas
python-dotenv

ruff
asyncio